Rework side-effect wiring
Side-effects now bypass the intent() layer, for intent():

- ought to deal with the *user* intent, not with the automated side
  effects
- is wired onto sources.DOM, and composing with another source has proven
  tough

The important thing is for main() to get a mutation_proposal$ to feed the
model with, no matter what the source of a specific proposal is. The
current implementation regroups several proposal sources (DOM, NAP) into
a single stream, as is expected.

This next-action-predicate layer, nap(), behaves just like intent() but
is simply wired to another source so made independent.